refactor: Extract models from view model to their own files.
